test app

+++ src/Providers/Przelewy24/Models/RefundTransactionRequest.cs
@@ -10,7 +10,7 @@ internal sealed class RefundTransactionRequest
 
 internal sealed class RefundItem
 {
-    [JsonPropertyName("orderId")]     public int OrderId { get; init; }
+    [JsonPropertyName("orderId")]     public long OrderId { get; init; }
     [JsonPropertyName("sessionId")]   public required string SessionId { get; init; }
     [JsonPropertyName("amount")]      public int Amount { get; init; }
